Move ColorItem and QColorEditor to params package
