+++ silx/gui/plot3d/SFViewParamTree.py
@@ -33,7 +33,6 @@ __license__ = "MIT"
 __date__ = "02/10/2017"
 
 import logging
-import sys
 
 import numpy
 
@@ -43,7 +42,7 @@ from silx.gui.plot.Colormap import Colormap
 from silx.gui.widgets.FloatEdit import FloatEdit
 
 from .ScalarFieldView import Isosurface
-from .params import SubjectItem
+from .params import SubjectItem, ColorItem, QColorEditor
 
 _logger = logging.getLogger(__name__)
 
@@ -54,34 +53,6 @@ class ModelColumns(object):
 
 
 # View settings ###############################################################
-
-class ColorItem(SubjectItem):
-    """color item."""
-    editable = True
-    persistent = True
-
-    def getEditor(self, parent, option, index):
-        editor = QColorEditor(parent)
-        editor.color = self.getColor()
-
-        # Wrapping call in lambda is a workaround for PySide with Python 3
-        editor.sigColorChanged.connect(
-            lambda color: self._editorSlot(color))
-        return editor
-
-    def _editorSlot(self, color):
-        self.setData(color, qt.Qt.EditRole)
-
-    def _pushData(self, value, role=qt.Qt.UserRole):
-        self.setColor(value)
-        return self.getColor()
-
-    def _pullData(self):
-        self.getColor()
-
-    def setColor(self, color):
-        """Override to implement actual color setter"""
-        pass
 
 
 class BackgroundColorItem(ColorItem):
@@ -610,67 +581,6 @@ class IsoSurfaceColorItem(SubjectItem):
     def _pushData(self, value, role=qt.Qt.UserRole):
         self.getSubject().setColor(value)
         return self.getSubject().getColor()
-
-
-class QColorEditor(qt.QWidget):
-    """
-    QColor editor.
-    """
-    sigColorChanged = qt.Signal(object)
-
-    color = property(lambda self: qt.QColor(self.__color))
-
-    @color.setter
-    def color(self, color):
-        self._setColor(color)
-        self.__previousColor = color
-
-    def __init__(self, *args, **kwargs):
-        super(QColorEditor, self).__init__(*args, **kwargs)
-        layout = qt.QHBoxLayout(self)
-        layout.setContentsMargins(0, 0, 0, 0)
-        button = qt.QToolButton()
-        icon = qt.QIcon(qt.QPixmap(32, 32))
-        button.setIcon(icon)
-        layout.addWidget(button)
-        button.clicked.connect(self.__showColorDialog)
-        layout.addStretch(1)
-
-        self.__color = None
-        self.__previousColor = None
-
-    def sizeHint(self):
-        return qt.QSize(0, 0)
-
-    def _setColor(self, qColor):
-        button = self.findChild(qt.QToolButton)
-        pixmap = qt.QPixmap(32, 32)
-        pixmap.fill(qColor)
-        button.setIcon(qt.QIcon(pixmap))
-        self.__color = qColor
-
-    def __showColorDialog(self):
-        dialog = qt.QColorDialog(parent=self)
-        if sys.platform == 'darwin':
-            # Use of native color dialog on macos might cause problems
-            dialog.setOption(qt.QColorDialog.DontUseNativeDialog, True)
-
-        self.__previousColor = self.__color
-        dialog.setAttribute(qt.Qt.WA_DeleteOnClose)
-        dialog.setModal(True)
-        dialog.currentColorChanged.connect(self.__colorChanged)
-        dialog.finished.connect(self.__dialogClosed)
-        dialog.show()
-
-    def __colorChanged(self, color):
-        self.__color = color
-        self._setColor(color)
-        self.sigColorChanged.emit(color)
-
-    def __dialogClosed(self, result):
-        if result == qt.QDialog.Rejected:
-            self.__colorChanged(self.__previousColor)
-        self.__previousColor = None
 
 
 class IsoSurfaceAlphaItem(SubjectItem):
