+++ silx/gui/plot3d/params/__init__.py
@@ -34,3 +34,4 @@ __date__ = "24/11/2017"
 
 
 from .SubjectItem import SubjectItem  # noqa
+from .color import ColorItem, QColorEditor  # noqa
